+++ frontend/packages/data-portal/app/components/InlineMetadata.tsx
@@ -1,9 +1,12 @@
-import { cns } from 'app/utils/cns'
+import { cns, cnsNoMerge } from 'app/utils/cns'
 
 export interface Metadata {
   key: string
   value: string
+
   uppercase?: boolean
+  keyClass?: string
+  valueClass?: string
 }
 
 export function InlineMetadata({
@@ -17,7 +20,7 @@ export function InlineMetadata({
 }) {
   const fieldsRender = (
     <ul className="list-none flex gap-sds-l items-baseline">
-      {fields.map(({ key, value, uppercase }) => (
+      {fields.map(({ key, value, uppercase, keyClass, valueClass }) => (
         <li
           className={cns(
             'flex flex-row items-baseline justify-left gap-sds-xxs',
@@ -26,24 +29,26 @@ export function InlineMetadata({
           key={key + value}
         >
           <span
-            className={cns(
+            className={cnsNoMerge(
               'font-semibold',
               subheader
-                ? 'text-sds-gray-600 tracking-sds-caps text-sds-caps-xxs leading-sds-caps-xxs'
+                ? 'text-sds-gray-600 tracking-sds-body-xxs text-sds-caps-xxs leading-sds-caps-xxs'
                 : 'text-sds-body-xxs leading-sds-body-xxs',
               uppercase && 'uppercase',
               subheader && uppercase && 'tracking-sds-caps',
+              keyClass,
             )}
           >
             {key}:
           </span>
 
           <span
-            className={
+            className={cnsNoMerge(
               subheader
                 ? 'text-sds-body-s leading-sds-body-s'
-                : 'text-sds-body-xxs leading-sds-body-xxs'
-            }
+                : 'text-sds-body-xxs leading-sds-body-xxs',
+              valueClass,
+            )}
           >
             {value}
           </span>