+++ frontend/packages/data-portal/e2e/constants.ts
@@ -15,6 +15,10 @@ export const E2E_CONFIG = merge(
   JSON.parse(process.env.E2E_CONFIG ?? '{}'),
 ) as E2EConfig
 
-export const BROWSE_DATASETS_URL = `${E2E_CONFIG.url}/browse-data/datasets`
-export const SINGLE_DATASET_URL = `${E2E_CONFIG.url}/datasets/${E2E_CONFIG.datasetId}`
-export const SINGLE_RUN_URL = `${E2E_CONFIG.url}/runs/${E2E_CONFIG.runId}`
+export const BROWSE_DATASETS_PATH = '/browse-data/datasets'
+export const SINGLE_DATASET_PATH = `/datasets/${E2E_CONFIG.datasetId}`
+export const SINGLE_RUN_PATH = `/runs/${E2E_CONFIG.runId}`
+
+export const BROWSE_DATASETS_URL = `${E2E_CONFIG.url}${BROWSE_DATASETS_PATH}`
+export const SINGLE_DATASET_URL = `${E2E_CONFIG.url}${SINGLE_DATASET_PATH}`
+export const SINGLE_RUN_URL = `${E2E_CONFIG.url}${SINGLE_RUN_PATH}`