+++ frontend/packages/data-portal/e2e/pageObjects/metadataDrawer/utils.ts
@@ -1,3 +1,5 @@
+import { ApolloClient, NormalizedCacheObject } from '@apollo/client'
+import { E2E_CONFIG, translations } from 'e2e/constants'
 import { DeepPartial } from 'utility-types'
 
 import {
@@ -6,15 +8,17 @@ import {
   Tiltseries,
   Tomograms,
 } from 'app/__generated__/graphql'
+import { getDatasetById } from 'app/graphql/getDatasetById.server'
+import { getRunById } from 'app/graphql/getRunById.server'
 import { isFiducial } from 'app/utils/tomograms'
 
-import { DrawerTestMetadata } from './types'
+import { DrawerTestData, DrawerTestMetadata } from './types'
 
-export function getBoolString(value?: boolean): string {
+function getBoolString(value?: boolean): string {
   return value ? 'True' : 'False'
 }
 
-export function getDatasetTestMetadata({
+function getDatasetTestMetadata({
   dataset,
   type,
 }: {
@@ -25,7 +29,6 @@ export function getDatasetTestMetadata({
     cellLineOrStrainName: dataset.cell_strain_name,
     cellName: dataset.cell_name,
     cellularComponent: dataset.cell_component_name,
-    citations: dataset.dataset_citations?.split(', ') ?? [],
     depositionDate: dataset.deposition_date,
     fundingAgency: dataset?.funding_sources?.map(
       (source) => source?.funding_agency_name ?? '',
@@ -51,7 +54,7 @@ export function getDatasetTestMetadata({
   }
 }
 
-export function getTiltSeriesTestMetadata({
+function getTiltSeriesTestMetadata({
   tiltSeries,
   type,
 }: {
@@ -88,7 +91,7 @@ export function getTiltSeriesTestMetadata({
   }
 }
 
-export function getTomogramTestMetadata(
+function getTomogramTestMetadata(
   tomogram: DeepPartial<Tomograms>,
 ): DrawerTestMetadata {
   return {
@@ -105,10 +108,13 @@ export function getTomogramTestMetadata(
   }
 }
 
-export function getAnnotationTestMetdata(
+function getAnnotationTestMetdata(
   annotation: DeepPartial<Annotations>,
 ): DrawerTestMetadata {
   const file = (annotation.files ?? []).at(0)
+
+  const getGroundTruthField = <T>(value: T) =>
+    annotation.ground_truth_status ? translations.notApplicable : value
 
   return {
     annotationId: annotation.id,
@@ -126,9 +132,89 @@ export function getAnnotationTestMetdata(
     objectShapeType: file?.shape_type,
     objectState: annotation.object_state,
     objectDescription: annotation.object_description,
+
+    // Ground truth annotations show N/A for precision and recall because they
+    // represent the correct or true labels for a dataset. Non ground truth
+    // annotations have these fields because they are compared against a ground
+    // truth annotation.
+    precision: getGroundTruthField(annotation.confidence_precision ?? '--'),
+    recall: getGroundTruthField(annotation.confidence_recall ?? '--'),
     groundTruthStatus: getBoolString(annotation.ground_truth_status),
-    groundTruthUsed: annotation.ground_truth_used,
-    precision: annotation.confidence_precision,
-    recall: annotation.confidence_recall,
   }
 }
+
+export async function getSingleDatasetTestMetadata(
+  client: ApolloClient<NormalizedCacheObject>,
+): Promise<DrawerTestData> {
+  const { data } = await getDatasetById({
+    client,
+    id: +E2E_CONFIG.datasetId,
+  })
+
+  const [dataset] = data.datasets
+  const [tiltSeries] = dataset.run_metadata[0].tiltseries
+
+  return {
+    title: dataset.title,
+    metadata: {
+      ...getDatasetTestMetadata({
+        dataset,
+        type: 'dataset',
+      }),
+
+      ...getTiltSeriesTestMetadata({
+        tiltSeries,
+        type: 'dataset',
+      }),
+    },
+  }
+}
+
+// #region Data Getters
+export async function getSingleRunTestMetadata(
+  client: ApolloClient<NormalizedCacheObject>,
+): Promise<DrawerTestData> {
+  const { data } = await getRunById({
+    client,
+    id: +E2E_CONFIG.runId,
+  })
+
+  const [run] = data.runs
+  const [tiltSeries] = run.tiltseries
+  const [tomogram] = run.tomogram_voxel_spacings[0].tomograms
+
+  return {
+    title: run.name,
+    metadata: {
+      ...getDatasetTestMetadata({
+        dataset: run.dataset,
+        type: 'run',
+      }),
+
+      ...getTiltSeriesTestMetadata({
+        tiltSeries,
+        type: 'run',
+      }),
+
+      ...getTomogramTestMetadata(tomogram),
+    },
+  }
+}
+
+export async function getAnnotationTestData(
+  client: ApolloClient<NormalizedCacheObject>,
+) {
+  const { data } = await getRunById({
+    client,
+    id: +E2E_CONFIG.runId,
+  })
+
+  const [run] = data.runs
+  const annotation = run.annotation_table[0].annotations[0]
+
+  return {
+    title: `${annotation.id} - ${annotation.object_name}`,
+    metadata: getAnnotationTestMetdata(annotation),
+  }
+}
+// #endregion Data Getters